requirements.txt
# ============================================
# CSV 数据分析系统 - 依赖包列表
# ============================================

# LLM API 客户端
zhipuai>=2.0.0

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# CLI 用户界面
rich>=13.0.0
prompt-toolkit>=3.0.0

# 配置文件管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0

# 工具库
python-dateutil>=2.8.0

# CSV数据分析系统

